fix double allocation on push_str that promotes from inline
Currently, when pushing onto an Inline SmartString, there will be a
double allocation (once in the `to_string` of the inlined string and
again when pushing the new string onto the end.

This patch reserves enough memory for the new string to fit so that
there is only one allocation.